// app/types/common.ts
export type PaginationParams = {
    page?: number;
    limit?: number;
    total?: number;
    totalPages?: number;
};


export interface PaginationMeta {
    page: number;
    limit: number;
    total: number;
    totalPages: number;
}

export interface PaginatedResponse<T> {
    data: T[];
    meta: PaginationMeta;
}